GPU Display: Split GPUDisplay.cxx in multiple files
