+++ GPU/GPUTracking/display/backend/GPUDisplayBackendVulkan.h
@@ -65,7 +65,7 @@ class GPUDisplayBackendVulkan : public GPUDisplayBackend
   unsigned int drawVertices(const vboList& v, const drawType t) override;
   void ActivateColor(std::array<float, 4>& color) override;
   void SetVSync(bool enable) override { mMustUpdateSwapChain = true; };
-  void setDepthBuffer() override{};
+  void setDepthBuffer() override {};
   bool backendNeedRedraw() override;
   int InitBackendA() override;
   void ExitBackendA() override;
