+++ GPU/GPUTracking/display/helpers/GPUDisplayBackendOpenGLMagneticField.cxx
@@ -38,7 +38,7 @@ using namespace GPUCA_NAMESPACE::gpu;
 #define GPUCA_BUILD_EVENT_DISPLAY_OPENGL
 #if !defined(GL_VERSION_4_5) || GL_VERSION_4_5 != 1
 #ifdef GPUCA_STANDALONE
-//#error Unsupported OpenGL version < 4.5
+// #error Unsupported OpenGL version < 4.5
 #elif defined(GPUCA_O2_LIB)
 #pragma message "Unsupported OpenGL version < 4.5, disabling standalone event display"
 #else
